+++ vulkano/src/command_buffer/commands/pipeline.rs
@@ -647,7 +647,8 @@ where
             let layout_binding =
                 &pipeline.layout().set_layouts()[set_num as usize].bindings()[&binding_num];
 
-            let check_buffer = |_index: u32, (_buffer, _range): &(Arc<dyn BufferAccess>, Range<DeviceSize>)| Ok(());
+            let check_buffer =
+                |_index: u32, (_buffer, _range): &(Arc<dyn BufferAccess>, Range<DeviceSize>)| Ok(());
 
             let check_buffer_view = |index: u32, buffer_view: &Arc<dyn BufferViewAbstract>| {
                 for desc_reqs in (binding_reqs.descriptors.get(&Some(index)).into_iter())
@@ -1574,7 +1575,7 @@ where
                 None => return Err(PipelineExecutionError::VertexBufferNotBound { binding_num }),
             };
 
-            let mut num_elements = vertex_buffer.size() as u64 / binding_desc.stride as u64;
+            let mut num_elements = vertex_buffer.size() / binding_desc.stride as u64;
 
             match binding_desc.input_rate {
                 VertexInputRate::Vertex => {
@@ -2146,13 +2147,13 @@ impl SyncCommandBufferBuilder {
 
             let descriptor_set_state = &descriptor_sets_state.descriptor_sets[&set];
 
-            match descriptor_set_state.resources()
-                .binding(binding)
-                .unwrap()
-            {
+            match descriptor_set_state.resources().binding(binding).unwrap() {
                 DescriptorBindingResources::None(_) => continue,
                 DescriptorBindingResources::Buffer(elements) => {
-                    if matches!(descriptor_type, DescriptorType::UniformBufferDynamic | DescriptorType::StorageBufferDynamic) {
+                    if matches!(
+                        descriptor_type,
+                        DescriptorType::UniformBufferDynamic | DescriptorType::StorageBufferDynamic
+                    ) {
                         let dynamic_offsets = descriptor_set_state.dynamic_offsets();
                         resources.extend(
                             (elements.iter().enumerate())
@@ -2163,7 +2164,8 @@ impl SyncCommandBufferBuilder {
                                         (
                                             index as u32,
                                             buffer.clone(),
-                                            dynamic_offset + range.start..dynamic_offset + range.end,
+                                            dynamic_offset + range.start
+                                                ..dynamic_offset + range.end,
                                         )
                                     })
                                 })
@@ -2174,11 +2176,7 @@ impl SyncCommandBufferBuilder {
                             (elements.iter().enumerate())
                                 .filter_map(|(index, element)| {
                                     element.as_ref().map(|(buffer, range)| {
-                                        (
-                                            index as u32,
-                                            buffer.clone(),
-                                            range.clone(),
-                                        )
+                                        (index as u32, buffer.clone(), range.clone())
                                     })
                                 })
                                 .flat_map(buffer_resource),
